Update to master

+++ euler/three.rs
@@ -1,7 +1,5 @@
 /// Find the sum of all the multiples of 3 or 5 below 1000.
 
-use std::iter::AdditiveIterator;
-
 struct PrimeFactorIterator {
     priv number: uint,
     priv current: uint,
@@ -19,10 +17,10 @@ impl PrimeFactorIterator {
 impl Iterator<uint> for PrimeFactorIterator {
     fn next(&mut self) -> Option<uint> {
         fn isprime(n: uint) -> bool {
-            range(2, (n as float).sqrt().ceil() as uint).all(|x| n % x != 0)
+            range(2, (n as f64).sqrt().ceil() as uint).all(|x| n % x != 0)
         }
 
-        while self.current < (self.number as float).sqrt().ceil() as uint  {
+        while self.current < (self.number as f64).sqrt().ceil() as uint  {
             self.current += 1;
             if (self.number % self.current == 0) & isprime(self.current) {
                 return Some(self.current);
